refactor(frontend): remove re-export shim, import ClarificationQuestionsCard directly
Update CreateAgent and EditAgent to import ClarificationQuestionsCard
from the shared copilot/components/ location directly. Delete the
re-export shim to comply with no-barrel-files guideline.
