--- autogpt_platform/frontend/src/app/(platform)/copilot/tools/CreateAgent/components/ClarificationQuestionsCard.tsx
@@ -1,1 +1,0 @@
-export { ClarificationQuestionsCard } from "../../../components/ClarificationQuestionsCard/ClarificationQuestionsCard";
